+++ src/delete.ts
@@ -158,28 +158,34 @@ export function deleteElement(
           fs.rmSync(listMetadataFilePath);
         }
 
+        let parentElementOfListStoredToDisk = null;
         // delete list from parent element
-        deleteChildFromParentElement(parentElement, elementPathInfo.keyName);
+        if (parentIsAnElement && fs.existsSync(parentFilePath)) {
+          deleteChildFromParentElement(parentElement, elementPathInfo.keyName);
 
-        // load parent element into memory for YdsResult object
-        const parentElementOfListContentsToStore = yaml.dump(parentElement);
-        fs.writeFileSync(
-          parentFilePath,
-          parentElementOfListContentsToStore,
-          "utf-8"
-        );
-        const parentElementOfListStoredToDisk = load(
-          workingDirectoryPath,
-          parentElementPath,
-          depth
-        ).element;
+          // load parent element into memory for YdsResult object
+          const parentElementOfListContentsToStore = yaml.dump(parentElement);
+          fs.writeFileSync(
+            parentFilePath,
+            parentElementOfListContentsToStore,
+            "utf-8"
+          );
+          parentElementOfListStoredToDisk = load(
+            workingDirectoryPath,
+            parentElementPath,
+            depth
+          ).element;
+
+          // return result of delete list operation
+          return new YdsResult(
+            true,
+            parentElementOfListStoredToDisk,
+            parentElementPath
+          );
+        }
 
         // return result of delete list operation
-        return new YdsResult(
-          true,
-          parentElementOfListStoredToDisk,
-          parentElementPath
-        );
+        return new YdsResult(true, null, parentElementPath);
       case ElementPathType.shortToSimple:
       case ElementPathType.hierarchicalToSimple:
         deleteChildFromParentElement(parentElement, elementPathInfo.keyName);
